pyproject.toml
[tool.poetry]
name = "gyt-cli"
version = "0.1.0"
description = "A Python CLI abstracting common Git interactions like Conventional Commit and Atlassian Smart Commit helpers"
authors = ["Nate Ferrell <nate@scrthq.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "gyt_cli", from = "." },
]

[tool.poetry.scripts]
gyt = "gyt_cli.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.12.0"}
GitPython = "^3.1.43"
PyYAML = "^6.0.1"
pydantic = "^2.6.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.3.0"
pytest-html = "^4.1.1"
pytest-black = "^0.3.12"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
